requirements.txt
telethon==1.36.0
mcp[cli]>=1.0.0
python-dotenv
uvicorn